Updated Test cases
